Improve NFT images loading time
`fetchHashTable` resorts to `MetadataProgram.getProgramAccounts` to get all accounts and that takes a lot of time to respond.
To improve the loading time a bit, we can request the NFT metadata in parallel (as opposed to sequentially). Instead of waiting for the first NFT's metadata to be loaded before requesting the second NFT's metadata (this creates a bottleneck considering we'll have 3 sequential requests), we request all the metadata at once and then wait until we get it all back to store all the images in the state in one go.